Add files via upload
I'm excited to share my project,"Calculator" it's a web-based  calculator built from the ground up using Html ,CSS and JavaScript .One of the key highlights of this project is it's sleek and user-friendly interface. I've put a lot of effort into crafting a design that not only looks great but also enhances the user experience.